+++ pyslvs/__init__.py
@@ -12,7 +12,7 @@ from .expression import get_vlinks, VJoint, VPoint, VLink, Coord
 from .bfgs import SolverSystem
 from .topo_config import t_config, EStack
 from .tinycadlib import (
-    pxy, ppp, plap, pllp, plpp, palp, vpoint_dof, expr_parser, expr_solving,
+    pxy, ppp, plap, pllp, plpp, palp, vpoint_dof, expr_solving,
 )
 # from .planar_linkage import (
 #     FMatch, norm_path, curvature, derivative, path_signature, cross_correlation,